Changes to Worker, now started as a Thread to be able to run Benchmark. UpdateDBListener which receives Message with Updates for inmemoryDB.
Car/Part aadditional informatiuon for testing and order.

Added some Quueues, now it works with 2 queues each time and selectors
(prefer cars before carBody for Painter) and HiPrio and normal Queue
for each task for the Workers.

Benchmark Initial Testing works now with NEW Workers since Threads can
be started and stopped, as required by assignment.

Added TestCase Object which handles the two testCaseTypes to see if all
tests were done.

+++ src/at/ac/sbc/carfactory/ui/AssignWorkPanel.java
@@ -118,7 +118,7 @@ public class AssignWorkPanel extends JDialog {
                         new NumberFormatter(percentDisplayFormat),
                         percentEditFormatter));
 
-		textFieldErrorRate.setValue(new Float(.15));
+		textFieldErrorRate.setValue(new Double(.15));
 		textFieldErrorRate.setColumns(5);
 		// textFieldErrorRate.addPropertyChangeListener("value", this);
